requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
alembic==1.13.1

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Email
imaplib2==3.6

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# Caching
redis==5.0.1
Flask-Caching==2.1.0

# Background Workers
celery==5.3.4
python-daemon==3.0.1

# ML/AI
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4

# URL Analysis
beautifulsoup4==4.12.2
lxml==4.9.3
tldextract==5.1.1

# Threat Intelligence
dnspython==2.4.2
whois==0.9.25

# Logging
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
validators==0.22.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0

# Production Server
gunicorn==21.2.0
waitress==2.1.2

# Development
watchdog==3.0.0
